requirements-new.txt
asgiref==3.7.2
black==24.2.0
boto3==1.34.45
botocore==1.34.45
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-storages==1.14.2
djangorestframework==3.14.0
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.35
idna==3.6
jmespath==1.0.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-memcached==1.62
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
virtualenv==20.25.0
